final_prog/PC/requirements.txt
flask
paho-mqtt
numpy
requests
git+https://github.com/jrialland/python-astar.git
websocket-server==0.4